Had to change customerID to int
I had no idea that javascript could not process Longs.

+++ src/main/java/com/example/demo/model/OrderCustomerMapper.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class OrderCustomerMapper {
-    Long customerID;
+    int customerID;
     List<Product> productsToAdd;
 }
